Conflicts:
	code/game/asteroid.dm
	code/modules/mining/monolith.dm

+++ code/game/machinery/bots/cleanbot.dm
@@ -40,6 +40,7 @@
 	var/next_dest_loc
 	radio_frequency = SERV_FREQ //Service
 	bot_type = CLEAN_BOT
+	model = "Cleanbot"
 
 /obj/machinery/bot/cleanbot/New()
 	..()
@@ -47,7 +48,7 @@
 	icon_state = "cleanbot[on]"
 
 	var/datum/job/janitor/J = new/datum/job/janitor
-	botcard.access = J.get_access()
+	botcard.access += J.get_access()
 	prev_access = botcard.access
 
 /obj/machinery/bot/cleanbot/turn_on()
@@ -109,10 +110,6 @@ text("<A href='?src=\ref[src];power=1'>[on ? "On" : "Off"]</A>"))
 			prints = !prints
 			get_targets()
 			updateUsrDialog()
-		if("freq")
-			var/freq = text2num(input("Select frequency for  navigation beacons", "Frequency", num2text(beacon_freq / 10))) * 10
-			if (freq > 0)
-				beacon_freq = freq
 			updateUsrDialog()
 
 /obj/machinery/bot/cleanbot/attackby(obj/item/weapon/W, mob/user as mob, params)
@@ -175,7 +172,7 @@ text("<A href='?src=\ref[src];power=1'>[on ? "On" : "Off"]</A>"))
 	if(target)
 		if(!path || path.len == 0) //No path, need a new one
 			//Try to produce a path to the target, and ignore airlocks to which it has access.
-			path = get_path_to(loc, target.loc, /turf/proc/AdjacentTurfsWithAccess, /turf/proc/Distance, 0, 30, id=botcard)
+			path = get_path_to(loc, target.loc, src, /turf/proc/Distance, 0, 30, id=botcard)
 			if (!bot_move(target))
 				add_to_ignore(target)
 				target = null
@@ -256,7 +253,8 @@ text("<A href='?src=\ref[src];power=1'>[on ? "On" : "Off"]</A>"))
 /obj/item/weapon/bucket_sensor/attackby(var/obj/item/W, mob/user as mob, params)
 	..()
 	if(istype(W, /obj/item/robot_parts/l_arm) || istype(W, /obj/item/robot_parts/r_arm))
-		user.drop_item()
+		if(!user.unEquip(W))
+			return
 		qdel(W)
 		var/turf/T = get_turf(loc)
 		var/obj/machinery/bot/cleanbot/A = new /obj/machinery/bot/cleanbot(T)
