+++ code/game/objects/items.dm
@@ -1,3 +1,5 @@
+var/global/image/fire_overlay = image("icon" = 'icons/effects/fire.dmi', "icon_state" = "fire")
+
 /obj/item
 	name = "item"
 	icon = 'icons/obj/items.dmi'
@@ -13,7 +15,7 @@
 	var/w_class = 3.0
 	var/slot_flags = 0		//This is used to determine on which slots an item can fit.
 	pass_flags = PASSTABLE
-	pressure_resistance = 5
+	pressure_resistance = 3
 	var/obj/item/master = null
 
 	var/heat_protection = 0 //flags which determine which body parts are protected from heat. Use the HEAD, CHEST, GROIN, etc. flags. See setup.dm
@@ -36,6 +38,7 @@
 	var/siemens_coefficient = 1 // for electrical admittance/conductance (electrocution checks and shit)
 	var/slowdown = 0 // How much clothing is slowing you down. Negative values speeds you up
 	var/list/armor = list(melee = 0, bullet = 0, laser = 0,energy = 0, bomb = 0, bio = 0, rad = 0)
+	var/armour_penetration = 0 //percentage of armour effectiveness to remove
 	var/list/allowed = null //suit storage stuff.
 	var/obj/item/device/uplink/hidden/hidden_uplink = null // All items can have an uplink hidden inside, just remember to add the triggers.
 	var/strip_delay = 40
@@ -51,6 +54,8 @@
 
 	var/suittoggled = 0
 	var/hooded = 0
+
+	/obj/item/mouse_drag_pointer = MOUSE_ACTIVE_POINTER //the icon to indicate this object is being dragged
 
 	//So items can have custom embedd values
 	//Because customisation is king
@@ -69,8 +74,14 @@
 		/obj/structure/closet,
 		/obj/item/weapon/storage,
 		/obj/structure/safe,
-		/obj/machinery/disposal
+		/obj/machinery/disposal,
+		/obj/machinery/r_n_d/destructive_analyzer,
+		/obj/machinery/r_n_d/experimentor,
+		/obj/machinery/autolathe
 	)
+
+	var/flags_cover = 0 //for flags such as GLASSESCOVERSEYES
+
 /obj/item/proc/check_allowed_items(atom/target, not_inside)
 	if((src in target) || ((!istype(target.loc, /turf)) && (!istype(target, /turf)) && (not_inside)) || is_type_in_list(target, can_be_placed_into))
 		return 0
@@ -148,6 +159,23 @@
 
 /obj/item/attack_hand(mob/user as mob)
 	if (!user) return
+
+	if(burn_state == 1)
+		var/mob/living/carbon/human/H = user
+		if(istype(H))
+			if(H.gloves && (H.gloves.max_heat_protection_temperature > 360))
+				extinguish()
+				user << "<span class='notice'>You put out the fire on [src].</span>"
+			else
+				user << "<span class='warning'>You burn your hand on [src]!</span>"
+				var/obj/item/organ/limb/affecting = H.get_organ("[user.hand ? "l" : "r" ]_arm")
+				if(affecting.take_damage( 0, 5 ))		// 5 burn damage
+					H.update_damage_overlays(0)
+				H.updatehealth()
+				return
+		else
+			extinguish()
+
 	if (istype(src.loc, /obj/item/weapon/storage))
 		//If the item is in a storage item, take it out
 		var/obj/item/weapon/storage/S = src.loc
@@ -189,10 +217,10 @@
 /obj/item/attack_alien(mob/user as mob)
 	var/mob/living/carbon/alien/A = user
 
-	if(!A.has_fine_manipulation || w_class >= 4)
+	if(!A.has_fine_manipulation)
 		if(src in A.contents) // To stop Aliens having items stuck in their pockets
 			A.unEquip(src)
-		user << "Your claws aren't capable of such fine manipulation."
+		user << "<span class='warning'>Your claws aren't capable of such fine manipulation!</span>"
 		return
 	attack_paw(A)
 
@@ -234,7 +262,7 @@
 					else if(success)
 						user << "<span class='notice'>You put some things [S.preposition] [S].</span>"
 					else
-						user << "<span class='notice'>You fail to pick anything up with [S].</span>"
+						user << "<span class='warning'>You fail to pick anything up with [S]!</span>"
 
 			else if(S.can_be_inserted(src))
 				S.handle_item_insertion(src)
@@ -288,7 +316,7 @@
 	set category = "Object"
 	set name = "Pick up"
 
-	if(!usr.canmove || usr.stat || usr.restrained() || !Adjacent(usr))
+	if(usr.stat || usr.restrained() || !Adjacent(usr) || usr.stunned || usr.weakened || usr.lying)
 		return
 
 	if(ishuman(usr) || ismonkey(usr))
@@ -301,7 +329,7 @@
 			usr << "\red You already have something in your hand."
 		*/
 	else
-		usr << "<span class='notice'>This mob type can't use this verb.</span>"
+		usr << "<span class='warning'>This mob type can't use this verb!</span>"
 
 //This proc is executed when someone clicks the on-screen UI button. To make the UI button show, set the 'action_button_name'.
 //The default action is attack_self().
@@ -317,26 +345,30 @@
 
 /obj/item/proc/eyestab(mob/living/carbon/M as mob, mob/living/carbon/user as mob)
 
-	var/mob/living/carbon/human/H = M
-	if(istype(H) && ( \
-			(H.head && H.head.flags & HEADCOVERSEYES) || \
-			(H.wear_mask && H.wear_mask.flags & MASKCOVERSEYES) || \
-			(H.glasses && H.glasses.flags & GLASSESCOVERSEYES) \
-		))
-		// you can't stab someone in the eyes wearing a mask!
-		user << "<span class='danger'>You're going to need to remove that mask/helmet/glasses first.</span>"
-		return
+	var/is_human_victim = 0
+	if(ishuman(M))
+		is_human_victim = 1
+		var/mob/living/carbon/human/H = M
+		if((H.head && H.head.flags_cover & HEADCOVERSEYES) || \
+			(H.wear_mask && H.wear_mask.flags_cover & MASKCOVERSEYES) || \
+			(H.glasses && H.glasses.flags_cover & GLASSESCOVERSEYES))
+			// you can't stab someone in the eyes wearing a mask!
+			user << "<span class='danger'>You're going to need to remove that mask/helmet/glasses first!</span>"
+			return
 
-	var/mob/living/carbon/monkey/Mo = M
-	if(istype(Mo) && ( \
-			(Mo.wear_mask && Mo.wear_mask.flags & MASKCOVERSEYES) \
-		))
-		// you can't stab someone in the eyes wearing a mask!
-		user << "<span class='danger'>You're going to need to remove that mask/helmet/glasses first.</span>"
-		return
+	if(ismonkey(M))
+		var/mob/living/carbon/monkey/Mo = M
+		if(Mo.wear_mask && Mo.wear_mask.flags_cover & MASKCOVERSEYES)
+			// you can't stab someone in the eyes wearing a mask!
+			user << "<span class='danger'>You're going to need to remove that mask/helmet/glasses first!</span>"
+			return
 
 	if(isalien(M))//Aliens don't have eyes./N     slimes also don't have eyes!
-		user << "<span class='danger'>You cannot locate any eyes on this creature!</span>"
+		user << "<span class='warning'>You cannot locate any eyes on this creature!</span>"
+		return
+
+	if(isbrain(M))
+		user << "<span class='danger'>You cannot locate any organic eyes on this brain!</span>"
 		return
 
 	add_logs(user, M, "attacked", object="[src.name]", addition="(INTENT: [uppertext(user.a_intent)])")
@@ -351,7 +383,7 @@
 			"<span class='danger'>[user] has stabbed themself in the eyes with [src]!</span>", \
 			"<span class='userdanger'>You stab yourself in the eyes with [src]!</span>" \
 		)
-	if(istype(M, /mob/living/carbon/human))
+	if(is_human_victim)
 		var/mob/living/carbon/human/U = M
 		var/obj/item/organ/limb/affecting = U.get_organ("head")
 		if(affecting.take_damage(7))
@@ -359,6 +391,7 @@
 
 	else
 		M.take_organ_damage(7)
+
 	M.eye_blurry += rand(3,4)
 	M.eye_stat += rand(2,4)
 	if (M.eye_stat >= 10)
@@ -368,8 +401,8 @@
 			M << "<span class='danger'>Your eyes start to bleed profusely!</span>"
 		if(prob(50))
 			if(M.stat != 2)
-				M << "<span class='danger'>You drop what you're holding and clutch at your eyes!</span>"
-				M.drop_item()
+				if(M.drop_item())
+					M << "<span class='danger'>You drop what you're holding and clutch at your eyes!</span>"
 			M.eye_blurry += 10
 			M.Paralyse(1)
 			M.Weaken(2)
@@ -394,30 +427,40 @@
 		transfer_blood = 0
 		bloody_hands_mob = null
 
-/obj/item/singularity_pull(S, current_size)
+// decay so that items are thrown when they get close, not all the time
+/obj/item/singularity_decay(var/obj/singularity/S, current_size, dist)
+	if(anchored || prob(40)) return ..()
 	spawn(0) //this is needed or multiple items will be thrown sequentially and not simultaneously
-		if(current_size >= STAGE_FOUR)
-			throw_at(S,14,3)
-		else ..()
+		throw_at(S,14,3)
 
-/obj/item/acid_act(var/acidpwr, var/toxpwr, var/acid_volume)
+/obj/item/acid_act(var/acidpwr, var/acid_volume)
 	. = 1
+	if(unacidable)
+		return
+
+	var/meltingpwr = acid_volume*acidpwr
+	var/melting_threshold = 100
+	if(meltingpwr <= melting_threshold) // so a single unit can't melt items. You need 5.1+ unit for fluoro and 10.1+ for sulphuric
+		return
+
 	for(var/V in armor)
 		if(armor[V] > 0)
 			.-- //it survives the acid...
 			break
-	if(.)
+	if(. && prob(min(meltingpwr/10,90))) //chance to melt depends on acid power and volume.
 		var/turf/T = get_turf(src)
 		if(T)
-			T.visible_message("<span class='danger'>[src] melts away!</span>")
 			var/obj/effect/decal/cleanable/molten_item/I = new (T)
 			I.pixel_x = rand(-16,16)
 			I.pixel_y = rand(-16,16)
 			I.desc = "Looks like this was \an [src] some time ago."
+		if(istype(src,/obj/item/weapon/storage))
+			var/obj/item/weapon/storage/S = src
+			S.do_quick_empty() //melted storage item drops its content.
 		qdel(src)
 	else
 		for(var/armour_value in armor) //but is weakened
-			armor[armour_value] = max(armor[armour_value]-acidpwr,0)
+			armor[armour_value] = max(armor[armour_value]-min(acidpwr,meltingpwr/10),0)
 		if(!findtext(desc, "it looks slightly melted...")) //it looks slightly melted... it looks slightly melted... it looks slightly melted... etc.
 			desc += " it looks slightly melted..." //needs a space at the start, formatting
 
@@ -428,7 +471,8 @@
 		if(istype(A, /mob/living/carbon/human))
 			var/mob/living/carbon/human/H = A
 			if(can_embed(src))
-				if(prob(embed_chance))
+				if(prob(embed_chance) && !(PIERCEIMMUNE in H.dna.species.specflags))
+					H.throw_alert("embeddedobject")
 					var/obj/item/organ/limb/L = pick(H.organs)
 					L.embedded_objects |= src
 					add_blood(H)//it embedded itself in you, of course it's bloody!
@@ -440,3 +484,12 @@
 	//Reset regardless of if we hit a human.
 	throw_speed = initial(throw_speed) //explosions change this.
 	..()
+
+/obj/item/proc/remove_item_from_storage(atom/newLoc) //please use this if you're going to snowflake an item out of a obj/item/weapon/storage
+	if(!newLoc)
+		return 0
+	if(istype(loc,/obj/item/weapon/storage))
+		var/obj/item/weapon/storage/S = loc
+		S.remove_from_storage(src,newLoc)
+		return 1
+	return 0
