+++ code/game/objects/items/weapons/RCD.dm
@@ -110,9 +110,10 @@ RCD
 	if(istype(W, /obj/item/weapon/rcd_ammo))
 		var/obj/item/weapon/rcd_ammo/R = W
 		if((matter + R.ammoamt) > max_matter)
-			user << "<span class='notice'>The RCD cant hold any more matter-units.</span>"
+			user << "<span class='warning'>The RCD can't hold any more matter-units!</span>"
 			return
-		user.drop_item()
+		if(!user.unEquip(W))
+			return
 		qdel(W)
 		matter += R.ammoamt
 		playsound(src.loc, 'sound/machines/click.ogg', 50, 1)
@@ -127,19 +128,19 @@ RCD
 	switch(mode)
 		if(1)
 			mode = 2
-			user << "<span class='notice'>Changed mode to 'Airlock'</span>"
+			user << "<span class='notice'>You change RCD's mode to 'Airlock'.</span>"
 			if(prob(20))
 				src.spark_system.start()
 			return
 		if(2)
 			mode = 3
-			user << "<span class='notice'>Changed mode to 'Deconstruct'</span>"
+			user << "<span class='notice'>You change RCD's mode to 'Deconstruct'.</span>"
 			if(prob(20))
 				src.spark_system.start()
 			return
 		if(3)
 			mode = 1
-			user << "<span class='notice'>Changed mode to 'Floor & Walls'</span>"
+			user << "<span class='notice'>You change RCD's mode to 'Floor & Walls'.</span>"
 			if(prob(20))
 				src.spark_system.start()
 			return
@@ -160,21 +161,23 @@ RCD
 	switch(mode)
 		if(1)
 			if(istype(A, /turf/space))
+				var/turf/space/S = A
 				if(useResource(1, user))
-					user << "Building Floor..."
+					user << "<span class='notice'>You start building floor...</span>"
 					activate()
-					A:ChangeTurf(/turf/simulated/floor/plating)
+					S.ChangeTurf(/turf/simulated/floor/plating)
 					return 1
 				return 0
 
 			if(istype(A, /turf/simulated/floor))
+				var/turf/simulated/floor/F = A
 				if(checkResource(3, user))
-					user << "Building Wall ..."
+					user << "<span class='notice'>You start building wall...</span>"
 					playsound(src.loc, 'sound/machines/click.ogg', 50, 1)
-					if(do_after(user, 20))
+					if(do_after(user, 20, target = A))
 						if(!useResource(3, user)) return 0
 						activate()
-						A:ChangeTurf(/turf/simulated/wall)
+						F.ChangeTurf(/turf/simulated/wall)
 						return 1
 				return 0
 
@@ -188,9 +191,9 @@ RCD
 							break
 
 					if(door_check)
-						user << "Building Airlock..."
+						user << "<span class='notice'>You start building airlock...</span>"
 						playsound(src.loc, 'sound/machines/click.ogg', 50, 1)
-						if(do_after(user, 50))
+						if(do_after(user, 50, target = A))
 							if(!useResource(10, user)) return 0
 							activate()
 							var/obj/machinery/door/airlock/T = new airlock_type( A )
@@ -202,40 +205,45 @@ RCD
 							return 1
 						return 0
 					else
-						user << "There is another door here!"
+						user << "<span class='warning'>There is another door here!</span>"
 						return 0
 				return 0
 
 		if(3)
 			if(istype(A, /turf/simulated/wall))
-				if(istype(A, /turf/simulated/wall/r_wall) && !canRwall)
+				var/turf/simulated/wall/W = A
+				if(istype(W, /turf/simulated/wall/r_wall) && !canRwall)
 					return 0
 				if(checkResource(5, user))
-					user << "Deconstructing Wall..."
+					user << "<span class='notice'>You start deconstructing wall...</span>"
 					playsound(src.loc, 'sound/machines/click.ogg', 50, 1)
-					if(do_after(user, 40))
+					if(do_after(user, 40, target = A))
 						if(!useResource(5, user)) return 0
 						activate()
-						A:ChangeTurf(/turf/simulated/floor/plating)
+						W.ChangeTurf(/turf/simulated/floor/plating)
 						return 1
 				return 0
 
 			if(istype(A, /turf/simulated/floor))
-				if(checkResource(5, user))
-					user << "Deconstructing Floor..."
+				var/turf/simulated/floor/F = A
+				if(istype(F, F.baseturf))
+					user << "<span class='notice'>You can't dig any deeper!</span>"
+					return 0
+				else if(checkResource(5, user))
+					user << "<span class='notice'>You start deconstructing floor...</span>"
 					playsound(src.loc, 'sound/machines/click.ogg', 50, 1)
-					if(do_after(user, 50))
+					if(do_after(user, 50, target = A))
 						if(!useResource(5, user)) return 0
 						activate()
-						A:ChangeTurf(/turf/space)
+						F.ChangeTurf(F.baseturf)
 						return 1
 				return 0
 
 			if(istype(A, /obj/machinery/door/airlock))
 				if(checkResource(20, user))
-					user << "Deconstructing Airlock..."
+					user << "<span class='notice'>You start deconstructing airlock...</span>"
 					playsound(src.loc, 'sound/machines/click.ogg', 50, 1)
-					if(do_after(user, 50))
+					if(do_after(user, 50, target = A))
 						if(!useResource(20, user)) return 0
 						activate()
 						qdel(A)
